# update AliyunMirror image prefix