--- a/pkg/minikube/bootstrapper/images/images_test.go
+++ b/pkg/minikube/bootstrapper/images/images_test.go
@@ -127,6 +127,46 @@ func TestGetLatestTag(t *testing.T) {
 	}
 }
 
+func TestEssentialsAliyunMirror(t *testing.T) {
+	var testCases = []struct {
+		version string
+		images  []string
+	}{
+
+		{"v1.21.0", strings.Split(strings.Trim(`
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-apiserver:v1.21.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-controller-manager:v1.21.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-scheduler:v1.21.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-proxy:v1.21.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/pause:3.4.1
+registry.cn-hangzhou.aliyuncs.com/google_containers/etcd:3.4.13-0
+registry.cn-hangzhou.aliyuncs.com/google_containers/coredns:v1.8.0
+`, "\n"), "\n")},
+		{"v1.22.0", strings.Split(strings.Trim(`
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-apiserver:v1.22.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-controller-manager:v1.22.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-scheduler:v1.22.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/kube-proxy:v1.22.0
+registry.cn-hangzhou.aliyuncs.com/google_containers/pause:3.5
+registry.cn-hangzhou.aliyuncs.com/google_containers/etcd:3.5.0-0
+registry.cn-hangzhou.aliyuncs.com/google_containers/coredns:v1.8.4
+`, "\n"), "\n")},
+	}
+	for _, tc := range testCases {
+		t.Run(tc.version, func(t *testing.T) {
+			v, err := semver.Make(strings.TrimPrefix(tc.version, "v"))
+			if err != nil {
+				t.Fatal(err)
+			}
+			want := tc.images
+			got := essentials("registry.cn-hangzhou.aliyuncs.com/google_containers", v)
+			if diff := cmp.Diff(want, got); diff != "" {
+				t.Errorf("images mismatch (-want +got):\n%s", diff)
+			}
+		})
+	}
+}
+
 func TestAuxiliary(t *testing.T) {
 	want := []string{
 		"gcr.io/k8s-minikube/storage-provisioner:" + version.GetStorageProvisionerVersion(),
@@ -142,6 +182,16 @@ func TestAuxiliaryMirror(t *testing.T) {
 		"test.mirror/k8s-minikube/storage-provisioner:" + version.GetStorageProvisionerVersion(),
 	}
 	got := auxiliary("test.mirror")
+	if diff := cmp.Diff(want, got); diff != "" {
+		t.Errorf("images mismatch (-want +got):\n%s", diff)
+	}
+}
+
+func TestAuxiliaryAliyunMirror(t *testing.T) {
+	want := []string{
+		"registry.cn-hangzhou.aliyuncs.com/google_containers/storage-provisioner:" + version.GetStorageProvisionerVersion(),
+	}
+	got := auxiliary("registry.cn-hangzhou.aliyuncs.com/google_containers")
 	if diff := cmp.Diff(want, got); diff != "" {
 		t.Errorf("images mismatch (-want +got):\n%s", diff)
 	}