--- a/pkg/minikube/bootstrapper/images/repo.go
+++ b/pkg/minikube/bootstrapper/images/repo.go
@@ -16,8 +16,6 @@ limitations under the License.
 
 package images
 
-import "path"
-
 // DefaultKubernetesRepo is the default Kubernetes repository
 const DefaultKubernetesRepo = "k8s.gcr.io"
 
@@ -27,12 +25,4 @@ func kubernetesRepo(mirror string) string {
 		return mirror
 	}
 	return DefaultKubernetesRepo
-}
-
-// minikubeRepo returns the official minikube repository, or an alternate
-func minikubeRepo(mirror string) string {
-	if mirror == "" {
-		mirror = "gcr.io"
-	}
-	return path.Join(mirror, "k8s-minikube")
 }