--- a/pkg/minikube/bootstrapper/images/repo_test.go
+++ b/pkg/minikube/bootstrapper/images/repo_test.go
@@ -44,27 +44,3 @@ func Test_kubernetesRepo(t *testing.T) {
 	}
 
 }
-
-func Test_minikubeRepo(t *testing.T) {
-	tests := []struct {
-		mirror string
-		want   string
-	}{
-		{
-			"",
-			"gcr.io/k8s-minikube",
-		},
-		{
-			"mirror.k8s.io",
-			"mirror.k8s.io/k8s-minikube",
-		},
-	}
-
-	for _, tc := range tests {
-		got := minikubeRepo(tc.mirror)
-		if !cmp.Equal(got, tc.want) {
-			t.Errorf("mirror miss match, want: %s, got: %s", tc.want, got)
-		}
-	}
-
-}