--- a/pkg/minikube/node/cache.go
+++ b/pkg/minikube/node/cache.go
@@ -261,7 +261,7 @@ func imagesInConfigFile() ([]string, error) {
 
 func updateKicImageRepo(imgName string, repo string) string {
 	image := strings.TrimPrefix(imgName, "gcr.io/")
-	if repo == "registry.cn-hangzhou.aliyuncs.com/google_containers" {
+	if repo == constants.AliyunMirror {
 		// for aliyun registry must strip namespace from image name, e.g.
 		//   registry.cn-hangzhou.aliyuncs.com/google_containers/k8s-minikube/kicbase:v0.0.25 will not work
 		//   registry.cn-hangzhou.aliyuncs.com/google_containers/kicbase:v0.0.25 does work