<!-- 🎉 Thank you for contributing to minikube! 🎉 Here are some hints to get your PR merged faster:

1. Your PR title will be included in the release notes, choose it carefully
2. If the PR fixes an issue, add "fixes #<issue number>" to the description.
3. If the PR is a user interface change, please include a "before" and "after" example.
4. If the PR is a large design change, please include an enhancement proposal:
https://github.com/kubernetes/minikube/tree/master/enhancements
-->

fix #13768 

The `coredns` image is cached under `.minikube/cache/images/amd64/registry.cn-hangzhou.aliyuncs.com/google_containers/coredns/coredns_v1.8.6`.

The right path should be `.minikube/cache/images/amd64/registry.cn-hangzhou.aliyuncs.com/google_containers/coredns_v1.8.6`

Could you kindly review this PR?