@zhan9san 2022-03-18
Hi @medyagh 

Only the cache directory layout is wrong. After the tar file is loaded to docker daemon inside minikube container, a wrong image name is set in minikube(docker in docker) docker daemon. So `kubeadm` would pull image from remote registry. That's why `minikube` doesn't fail if host have internet access, but it would not work in offline usage.

## Before

1. cache directory layout

    ```
    vagrant@vagrant:~$ tree .minikube/cache/
    .minikube/cache/
    ├── images
    │   └── amd64
    │       └── registry.cn-hangzhou.aliyuncs.com
    │           └── google_containers
    │               ├── coredns
    │               │   └── coredns_v1.8.6
    │               ├── etcd_3.5.1-0
    │               ├── k8s-minikube
    │               │   └── storage-provisioner_v5
    │               ├── kube-apiserver_v1.23.3
    │               ├── kube-controller-manager_v1.23.3
    │               ├── kube-proxy_v1.23.3
    │               ├── kubernetesui
    │               │   ├── dashboard_v2.3.1
    │               │   └── metrics-scraper_v1.0.7
    │               ├── kube-scheduler_v1.23.3
    │               └── pause_3.6
    ├── kic
    │   └── amd64
    │       └── kicbase_v0.0.30.tar
    ├── linux
    │   └── amd64
    │       └── v1.23.3
    │           ├── kubeadm
    │           ├── kubectl
    │           └── kubelet
    └── preloaded-tarball
        └── preloaded-images-k8s-v17-v1.23.3-docker-overlay2-amd64.tar.lz4
    
    14 directories, 14 files
    ```

2. images digest
    
    As we can see, there are two coredns images and two storage-provisioner images.
    
    Both `registry.cn-hangzhou.aliyuncs.com/google_containers/coredns/coredns` and `registry.cn-hangzhou.aliyuncs.com/google_containers/k8s-minikube/storage-provisioner` are are loaded from cache, so the digest is lost.
    
    Why does other two exist?
    It is because the cache directory layout is wrong, which results in the wrong imaged named loaded to daemon.
    When kubeadm doesn't find the right one, it pulls the required image from remote registry, that's why the other two contain digest.
    
    ```shell
    $ minikube ssh
    Last login: Fri Mar 18 02:00:01 2022 from 192.168.49.1
    docker@minikube:~$ docker images --digests
    REPOSITORY                                                                             TAG       DIGEST                                                                    IMAGE ID       CREATED         SIZE
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-apiserver                     v1.23.3   <none>                                                                    f40be0088a83   7 weeks ago     135MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-proxy                         v1.23.3   <none>                                                                    9b7cc9982109   7 weeks ago     112MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-scheduler                     v1.23.3   <none>                                                                    99a3486be4f2   7 weeks ago     53.5MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-controller-manager            v1.23.3   <none>                                                                    b07520cd7ab7   7 weeks ago     125MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/etcd                               3.5.1-0   <none>                                                                    25f8c7f3da61   4 months ago    293MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/coredns                            v1.8.6    sha256:5b6ec0d6de9baaf3e92d0f66cd96a25b9edbce8716f5f15dcd1a616b3abd590e   a4ca41631cc7   5 months ago    46.8MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/coredns/coredns                    v1.8.6    <none>                                                                    a4ca41631cc7   5 months ago    46.8MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/pause                              3.6       <none>                                                                    6270bb605e12   6 months ago    683kB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kubernetesui/dashboard             v2.3.1    <none>                                                                    e1482a24335a   9 months ago    220MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kubernetesui/metrics-scraper       v1.0.7    <none>                                                                    7801cfc6d5c0   9 months ago    34.4MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/k8s-minikube/storage-provisioner   v5        <none>                                                                    6e38f40d628d   11 months ago   31.5MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/storage-provisioner                v5        sha256:18eb69d1418e854ad5a19e399310e52808a8321e4c441c1dddad8977a0d7a944   6e38f40d628d   11 months ago   31.5MB
    ```

## After

    1. cache directory layout
    
    ```
    $ tree ~/.minikube/cache/
    /home/vagrant/.minikube/cache/
    ├── images
    │   └── amd64
    │       └── registry.cn-hangzhou.aliyuncs.com
    │           └── google_containers
    │               ├── coredns_v1.8.6
    │               ├── dashboard_v2.3.1
    │               ├── etcd_3.5.1-0
    │               ├── kube-apiserver_v1.23.3
    │               ├── kube-controller-manager_v1.23.3
    │               ├── kube-proxy_v1.23.3
    │               ├── kube-scheduler_v1.23.3
    │               ├── metrics-scraper_v1.0.7
    │               ├── pause_3.6
    │               └── storage-provisioner_v5
    ├── kic
    │   └── amd64
    │       └── kicbase_v0.0.30.tar
    └── linux
        └── amd64
            └── v1.23.3
                ├── kubeadm
                ├── kubectl
                └── kubelet
    
    9 directories, 14 files
    ```

2. image digest

    ```
    vagrant@vagrant:~/bin$ minikube ssh
    docker@minikube:~$ docker images --digests
    REPOSITORY                                                                    TAG       DIGEST    IMAGE ID       CREATED         SIZE
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-apiserver            v1.23.3   <none>    f40be0088a83   7 weeks ago     135MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-controller-manager   v1.23.3   <none>    b07520cd7ab7   7 weeks ago     125MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-scheduler            v1.23.3   <none>    99a3486be4f2   7 weeks ago     53.5MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/kube-proxy                v1.23.3   <none>    9b7cc9982109   7 weeks ago     112MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/etcd                      3.5.1-0   <none>    25f8c7f3da61   4 months ago    293MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/coredns                   v1.8.6    <none>    a4ca41631cc7   5 months ago    46.8MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/pause                     3.6       <none>    6270bb605e12   6 months ago    683kB
    registry.cn-hangzhou.aliyuncs.com/google_containers/dashboard                 v2.3.1    <none>    e1482a24335a   9 months ago    220MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/metrics-scraper           v1.0.7    <none>    7801cfc6d5c0   9 months ago    34.4MB
    registry.cn-hangzhou.aliyuncs.com/google_containers/storage-provisioner       v5        <none>    6e38f40d628d   11 months ago   31.5MB
    ```

#### Before detailed log

```
vagrant@vagrant:~$ minikube start --driver=docker --image-mirror-country=cn --base-image=registry.cn-hangzhou.aliyuncs.com/google_containers/kicbase:v0.0.30 -v=8 --alsologtostderr
I0318 01:32:22.787992    3570 out.go:297] Setting OutFile to fd 1 ...
I0318 01:32:22.788231    3570 out.go:349] isatty.IsTerminal(1) = true
I0318 01:32:22.788244    3570 out.go:310] Setting ErrFile to fd 2...
I0318 01:32:22.788375    3570 out.go:349] isatty.IsTerminal(2) = true
I0318 01:32:22.788783    3570 root.go:315] Updating PATH: /home/vagrant/.minikube/bin
I0318 01:32:22.789231    3570 out.go:304] Setting JSON to false
I0318 01:32:22.790098    3570 start.go:112] hostinfo: {"hostname":"vagrant","uptime":2076,"bootTime":1647565067,"procs":109,"os":"linux","platform":"ubuntu","platformFamily":"debian","platformVersion":"21.04","kernelVersion":"5.11.0-38-generic","kernelArch":"x86_64","virtualizationSystem":"vbox","virtualizationRole":"guest","hostId":"185f9f8e-8dc3-4a84-b88f-64348f5ba85d"}
I0318 01:32:22.790169    3570 start.go:122] virtualization: vbox guest
I0318 01:32:22.792916    3570 out.go:176] 😄  minikube v1.25.2 on Ubuntu 21.04 (vbox/amd64)
😄  minikube v1.25.2 on Ubuntu 21.04 (vbox/amd64)
I0318 01:32:22.793835    3570 config.go:176] Loaded profile config "minikube": Driver=docker, ContainerRuntime=docker, KubernetesVersion=v1.23.3
W0318 01:32:22.794417    3570 start.go:706] api.Load failed for minikube: filestore "minikube": Docker machine "minikube" does not exist. Use "docker-machine ls" to list machines. Use "docker-machine create" to add a new one.
I0318 01:32:22.794706    3570 driver.go:344] Setting default libvirt URI to qemu:///system
W0318 01:32:22.794926    3570 start.go:706] api.Load failed for minikube: filestore "minikube": Docker machine "minikube" does not exist. Use "docker-machine ls" to list machines. Use "docker-machine create" to add a new one.
W0318 01:32:22.794146    3570 preload.go:295] Failed to list preload files: open /home/vagrant/.minikube/cache/preloaded-tarball: no such file or directory
I0318 01:32:22.794230    3570 notify.go:193] Checking for updates...
I0318 01:32:22.829333    3570 docker.go:132] docker version: linux-20.10.13
I0318 01:32:22.829526    3570 cli_runner.go:133] Run: docker system info --format "{{json .}}"
I0318 01:32:22.938894    3570 info.go:263] docker info: {ID:Y7GB:RDQJ:BDXD:DTAY:MYEB:KXFZ:U743:GBCV:EEZD:3YJF:SDJ3:JIUW Containers:0 ContainersRunning:0 ContainersPaused:0 ContainersStopped:0 Images:1 Driver:overlay2 DriverStatus:[[Backing Filesystem extfs] [Supports d_type true] [Native Overlay Diff true] [userxattr false]] SystemStatus:<nil> Plugins:{Volume:[local] Network:[bridge host ipvlan macvlan null overlay] Authorization:<nil> Log:[awslogs fluentd gcplogs gelf journald json-file local logentries splunk syslog]} MemoryLimit:true SwapLimit:true KernelMemory:true KernelMemoryTCP:true CPUCfsPeriod:true CPUCfsQuota:true CPUShares:true CPUSet:true PidsLimit:true IPv4Forwarding:true BridgeNfIptables:true BridgeNfIP6Tables:true Debug:false NFd:25 OomKillDisable:true NGoroutines:34 SystemTime:2022-03-18 01:32:22.859104376 +0000 UTC LoggingDriver:json-file CgroupDriver:cgroupfs NEventsListener:0 KernelVersion:5.11.0-38-generic OperatingSystem:Ubuntu 21.04 OSType:linux Architecture:x86_64 IndexServerAddress:https://index.docker.io/v1/ RegistryConfig:{AllowNondistributableArtifactsCIDRs:[] AllowNondistributableArtifactsHostnames:[] InsecureRegistryCIDRs:[127.0.0.0/8] IndexConfigs:{DockerIo:{Name:docker.io Mirrors:[] Secure:true Official:true}} Mirrors:[]} NCPU:2 MemTotal:2074591232 GenericResources:<nil> DockerRootDir:/var/lib/docker HTTPProxy: HTTPSProxy: NoProxy: Name:vagrant Labels:[] ExperimentalBuild:false ServerVersion:20.10.13 ClusterStore: ClusterAdvertise: Runtimes:{Runc:{Path:runc}} DefaultRuntime:runc Swarm:{NodeID: NodeAddr: LocalNodeState:inactive ControlAvailable:false Error: RemoteManagers:<nil>} LiveRestoreEnabled:false Isolation: InitBinary:docker-init ContainerdCommit:{ID:2a1d4dbdb2a1030dc5b01e96fb110a9d9f150ecc Expected:2a1d4dbdb2a1030dc5b01e96fb110a9d9f150ecc} RuncCommit:{ID:v1.0.3-0-gf46b6ba Expected:v1.0.3-0-gf46b6ba} InitCommit:{ID:de40ad0 Expected:de40ad0} SecurityOptions:[name=apparmor name=seccomp,profile=default] ProductLicense: Warnings:<nil> ServerErrors:[] ClientInfo:{Debug:false Plugins:[map[Experimental:true Name:app Path:/usr/libexec/docker/cli-plugins/docker-app SchemaVersion:0.1.0 ShortDescription:Docker App Vendor:Docker Inc. Version:v0.9.1-beta3] map[Name:buildx Path:/usr/libexec/docker/cli-plugins/docker-buildx SchemaVersion:0.1.0 ShortDescription:Docker Buildx Vendor:Docker Inc. Version:v0.8.0-docker] map[Name:scan Path:/usr/libexec/docker/cli-plugins/docker-scan SchemaVersion:0.1.0 ShortDescription:Docker Scan Vendor:Docker Inc. Version:v0.17.0]] Warnings:<nil>}}
I0318 01:32:22.938987    3570 docker.go:237] overlay module found
I0318 01:32:22.941695    3570 out.go:176] ✨  Using the docker driver based on existing profile
✨  Using the docker driver based on existing profile
I0318 01:32:22.941734    3570 start.go:281] selected driver: docker
I0318 01:32:22.941739    3570 start.go:798] validating driver "docker" against &{Name:minikube KeepContext:false EmbedCerts:false MinikubeISO: KicBaseImage:registry.cn-hangzhou.aliyuncs.com/google_containers/kicbase:v0.0.30 Memory:1978 CPUs:2 DiskSize:20000 VMDriver: Driver:docker HyperkitVpnKitSock: HyperkitVSockPorts:[] DockerEnv:[] ContainerVolumeMounts:[] InsecureRegistry:[] RegistryMirror:[] HostOnlyCIDR:192.168.59.1/24 HypervVirtualSwitch: HypervUseExternalSwitch:false HypervExternalAdapter: KVMNetwork:default KVMQemuURI:qemu:///system KVMGPU:false KVMHidden:false KVMNUMACount:1 DockerOpt:[] DisableDriverMounts:false NFSShare:[] NFSSharesRoot:/nfsshares UUID: NoVTXCheck:false DNSProxy:false HostDNSResolver:true HostOnlyNicType:virtio NatNicType:virtio SSHIPAddress: SSHUser:root SSHKey: SSHPort:22 KubernetesConfig:{KubernetesVersion:v1.23.3 ClusterName:minikube Namespace:default APIServerName:minikubeCA APIServerNames:[] APIServerIPs:[] DNSDomain:cluster.local ContainerRuntime:docker CRISocket: NetworkPlugin: FeatureGates: ServiceCIDR:10.96.0.0/12 ImageRepository:registry.cn-hangzhou.aliyuncs.com/google_containers LoadBalancerStartIP: LoadBalancerEndIP: CustomIngressCert: ExtraOptions:[{Component:kubelet Key:housekeeping-interval Value:5m}] ShouldLoadCachedImages:true EnableDefaultCNI:false CNI: NodeIP: NodePort:8443 NodeName:} Nodes:[{Name: IP: Port:8443 KubernetesVersion:v1.23.3 ContainerRuntime:docker ControlPlane:true Worker:true}] Addons:map[] CustomAddonImages:map[] CustomAddonRegistries:map[] VerifyComponents:map[apiserver:true system_pods:true] StartHostTimeout:6m0s ScheduledStop:<nil> ExposedPorts:[] ListenAddress: Network: MultiNodeRequested:false ExtraDisks:0 CertExpiration:26280h0m0s Mount:false MountString:/home/vagrant:/minikube-host Mount9PVersion:9p2000.L MountGID:docker MountIP: MountMSize:262144 MountOptions:[] MountPort:0 MountType:9p MountUID:docker BinaryMirror: DisableOptimizations:false}
I0318 01:32:22.941810    3570 start.go:809] status for docker: {Installed:true Healthy:true Running:false NeedsImprovement:false Error:<nil> Reason: Fix: Doc: Version:}
I0318 01:32:22.941874    3570 cli_runner.go:133] Run: docker system info --format "{{json .}}"
I0318 01:32:23.166062    3570 info.go:263] docker info: {ID:Y7GB:RDQJ:BDXD:DTAY:MYEB:KXFZ:U743:GBCV:EEZD:3YJF:SDJ3:JIUW Containers:0 ContainersRunning:0 ContainersPaused:0 ContainersStopped:0 Images:1 Driver:overlay2 DriverStatus:[[Backing Filesystem extfs] [Supports d_type true] [Native Overlay Diff true] [userxattr false]] SystemStatus:<nil> Plugins:{Volume:[local] Network:[bridge host ipvlan macvlan null overlay] Authorization:<nil> Log:[awslogs fluentd gcplogs gelf journald json-file local logentries splunk syslog]} MemoryLimit:true SwapLimit:true KernelMemory:true KernelMemoryTCP:true CPUCfsPeriod:true CPUCfsQuota:true CPUShares:true CPUSet:true PidsLimit:true IPv4Forwarding:true BridgeNfIptables:true BridgeNfIP6Tables:true Debug:false NFd:25 OomKillDisable:true NGoroutines:34 SystemTime:2022-03-18 01:32:22.967376108 +0000 UTC LoggingDriver:json-file CgroupDriver:cgroupfs NEventsListener:0 KernelVersion:5.11.0-38-generic OperatingSystem:Ubuntu 21.04 OSType:linux Architecture:x86_64 IndexServerAddress:https://index.docker.io/v1/ RegistryConfig:{AllowNondistributableArtifactsCIDRs:[] AllowNondistributableArtifactsHostnames:[] InsecureRegistryCIDRs:[127.0.0.0/8] IndexConfigs:{DockerIo:{Name:docker.io Mirrors:[] Secure:true Official:true}} Mirrors:[]} NCPU:2 MemTotal:2074591232 GenericResources:<nil> DockerRootDir:/var/lib/docker HTTPProxy: HTTPSProxy: NoProxy: Name:vagrant Labels:[] ExperimentalBuild:false ServerVersion:20.10.13 ClusterStore: ClusterAdvertise: Runtimes:{Runc:{Path:runc}} DefaultRuntime:runc Swarm:{NodeID: NodeAddr: LocalNodeState:inactive ControlAvailable:false Error: RemoteManagers:<nil>} LiveRestoreEnabled:false Isolation: InitBinary:docker-init ContainerdCommit:{ID:2a1d4dbdb2a1030dc5b01e96fb110a9d9f150ecc Expected:2a1d4dbdb2a1030dc5b01e96fb110a9d9f150ecc} RuncCommit:{ID:v1.0.3-0-gf46b6ba Expected:v1.0.3-0-gf46b6ba} InitCommit:{ID:de40ad0 Expected:de40ad0} SecurityOptions:[name=apparmor name=seccomp,profile=default] ProductLicense: Warnings:<nil> ServerErrors:[] ClientInfo:{Debug:false Plugins:[map[Experimental:true Name:app Path:/usr/libexec/docker/cli-plugins/docker-app SchemaVersion:0.1.0 ShortDescription:Docker App Vendor:Docker Inc. Version:v0.9.1-beta3] map[Name:buildx Path:/usr/libexec/docker/cli-plugins/docker-buildx SchemaVersion:0.1.0 ShortDescription:Docker Buildx Vendor:Docker Inc. Version:v0.8.0-docker] map[Name:scan Path:/usr/libexec/docker/cli-plugins/docker-scan SchemaVersion:0.1.0 ShortDescription:Docker Scan Vendor:Docker Inc. Version:v0.17.0]] Warnings:<nil>}}
I0318 01:32:23.169109    3570 out.go:176]

W0318 01:32:23.169468    3570 out.go:241] 🧯  The requested memory allocation of 1978MiB does not leave room for system overhead (total system memory: 1978MiB). You may face stability issues.
🧯  The requested memory allocation of 1978MiB does not leave room for system overhead (total system memory: 1978MiB). You may face stability issues.
W0318 01:32:23.170122    3570 out.go:241] 💡  Suggestion: Start minikube with less memory allocated: 'minikube start --memory=1978mb'
💡  Suggestion: Start minikube with less memory allocated: 'minikube start --memory=1978mb'
I0318 01:32:23.172901    3570 out.go:176]

I0318 01:32:23.173102    3570 cni.go:93] Creating CNI manager for ""
I0318 01:32:23.173209    3570 cni.go:167] CNI unnecessary in this configuration, recommending no CNI
I0318 01:32:23.173312    3570 start_flags.go:302] config:
…